Colorblind friendly palette from dittoSeq

Just to get the palette without having to install all those dependencies of dittoSeq.

Usage

ditto_colors

Format

A character vector of hex colors of the palette. There are 40 colors.

Source

The dittoSeq package.